[test] move api tests to tests/it
