+++ query/src/api/mod.rs
@@ -17,18 +17,14 @@
 pub use http_service::HttpService;
 pub use rpc::BroadcastAction;
 pub use rpc::CancelAction;
+pub use rpc::DatabendQueryFlightDispatcher;
 pub use rpc::FlightAction;
 pub use rpc::FlightClient;
 pub use rpc::FlightTicket;
 pub use rpc::ShuffleAction;
 pub use rpc_service::RpcService;
 
-mod http;
+pub mod http;
 mod http_service;
-mod rpc;
+pub mod rpc;
 mod rpc_service;
-
-#[cfg(test)]
-mod http_service_test;
-#[cfg(test)]
-mod rpc_service_test;
