+++ query/tests/it/api/http/mod.rs
@@ -12,7 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod cluster;
-pub mod config;
-pub mod health;
-pub mod logs;
+mod v1;
