+++ query/tests/it/api/http/v1/cluster.rs
@@ -14,7 +14,10 @@
 
 use common_base::tokio;
 use common_exception::Result;
+use common_meta_types::NodeInfo;
+use databend_query::api::http::v1::cluster::*;
 use poem::get;
+use poem::http::header;
 use poem::http::Method;
 use poem::http::StatusCode;
 use poem::http::Uri;
@@ -24,27 +27,32 @@ use poem::Request;
 use poem::Route;
 use pretty_assertions::assert_eq;
 
-use crate::api::http::v1::logs::logs_handler;
 use crate::tests::SessionManagerBuilder;
 
 #[tokio::test]
-async fn test_logs() -> Result<()> {
+async fn test_cluster() -> Result<()> {
     let sessions = SessionManagerBuilder::create().build()?;
-
-    let test_router = Route::new()
-        .at("/v1/logs", get(logs_handler))
+    let cluster_router = Route::new()
+        .at("/v1/cluster/list", get(cluster_list_handler))
         .data(sessions);
+
+    // List Node
     {
-        let response = test_router
+        let response = cluster_router
             .call(
                 Request::builder()
-                    .uri(Uri::from_static("/v1/logs"))
+                    .uri(Uri::from_static("/v1/cluster/list"))
+                    .header(header::CONTENT_TYPE, "application/json")
                     .method(Method::GET)
                     .finish(),
             )
             .await;
-
         assert_eq!(response.status(), StatusCode::OK);
+
+        let body = response.into_body().into_vec().await.unwrap();
+        let nodes = serde_json::from_str::<Vec<NodeInfo>>(&String::from_utf8_lossy(&body))?;
+        assert_eq!(nodes.len(), 1);
     }
+
     Ok(())
 }
