+++ query/tests/it/api/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod cluster;
-pub mod config;
-pub mod health;
-pub mod logs;
+mod http;
+mod http_service;
+mod rpc;
+mod rpc_service;
