+++ query/tests/it/main.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod api;
 mod servers;
 mod sessions;
 mod sql;
